only install once
Former-commit-id: 8cf651584ef1a2ff097cbd298b5d274288f3de8d

+++ test/singularity_run.sh
@@ -9,9 +9,12 @@ ARGS=${ARRAY[@]:4:$LEN}
 CODEDIR=$(dirname $WORKDIR)
 export PYTHONPATH=$WORKDIR
 
-python3 -m pip install --upgrade "pip<=21.0"
-python3 -m pip install --use-deprecated=legacy-resolver --upgrade-strategy=only-if-needed --no-cache-dir -U $CODEDIR
-python3 -m pip install --use-deprecated=legacy-resolver -U coverage
+if [ ! -f ~/.install_complete ]; then
+  python3 -m pip install --upgrade "pip<=21.0"
+  python3 -m pip install --use-deprecated=legacy-resolver --upgrade-strategy=only-if-needed --no-cache-dir -U $CODEDIR
+  python3 -m pip install --use-deprecated=legacy-resolver -U coverage
+  touch ~/.install_complete
+fi
 
 cd ${CODEDIR}
 python3 -m coverage run --rcfile=${WORKDIR}/.coveragerc --source=openproblems ${WORKDIR}/${SCRIPT} $TASKNAME $FUN ${ARGS}
